UI Change and Documentation Update
