--- docs/Discordsetup.txt
@@ -1,29 +1,0 @@
-The discord bot has a few requirements. Most of which are set up in the Github Secrets and a custom Discord server. 
-
-Firstly: 
-Create a new discord server if you don't have one already. 
-To copy the discord server id, enable "Developer Mode" in your settings. Then simply right click the server icon and scroll down to "Copy ID"
-Enter the server ID into the environment variable DISCORD_CLIENT_ID located in the docker-compose.yaml file. This is where the bot grabs the specific guild information. For reference, discord refers to their servers as guilds, they are basically the same thing. 
-
-Make a few custom roles in your discord server with titles CIS 115, CIS 200, CIS 300. Copy the role IDs and put them in server/configs/discordRoleIDs.
-This makes it easier to convert from the queried courses in the database into usable discord role IDs.
-
-For the discord bot. You must go to the discord developer portal and create a new application. 
-To create the URL for the bot, in the OAuth2 page select Bot and then Manage Roles, then guild install. The link generated from this will allow you to add the bot to your server. 
-
-You will need the following below as discord secrets in your Github secrets. All of which can be found in the application you created. 
-
-DISCORD_CLIENT_ID 
-DISCORD_CLIENT_SECRET
-DISCORD_SECRET
-
-Whoever creates the repo must make these secrets. Navigate to Settings -> Security -> Secrets and variables -> Codespaces. 
-Make a New repository secret with the exact format as above and add the values grabbed from the developer application.  
-
-Follow the steps here in the README https://github.com/JustinBeckwith/linked-role-bot to set up the linked roles.
-Glitch is not needed as the files are located in server/linked-roles/linked_roles_glitch. 
-In the tutorial, using the UUID as COOKIE_SECRET
-
-
-To test the automatic roles in the admin page, you must have seeding of a test user linked with a discord account. Make sure they have
-the classes that are seeded with the roles.
