+++ server/linkedRoles/Linked_Roles_Glitch/src/config.js
@@ -12,7 +12,7 @@ const config = {
   DISCORD_TOKEN: 'process.env.DISCORD_SECRET',
   DISCORD_CLIENT_ID: 'process.env.DISCORD_CLIENT_ID',
   DISCORD_CLIENT_SECRET: 'process.env.DISCORD_CLIENT_SECRET',
-  DISCORD_REDIRECT_URI: '${process.env.GITHUB_CODESPACE_URL}/discord-oauth-callback', //might need to be updated in developer portal
+  DISCORD_REDIRECT_URI: '${process.env.GITHUB_CODESPACE_URL}/discord-oauth-callback', //Must be updated in developer portal
   COOKIE_SECRET: 'process.env.COOKIE_SECRET',
 };
 export default config;
