+++ server/linkedRoles/Linked_Roles_Glitch/src/register.js
@@ -9,7 +9,7 @@ const url = `https://discord.com/api/v10/applications/${config.DISCORD_CLIENT_ID
 // supported types: number_lt=1, number_gt=2, number_eq=3 number_neq=4, datetime_lt=5, datetime_gt=6, boolean_eq=7
 const body = [
   {
-    key: 'verified_test3',
+    key: 'verified',
     name: 'Student Portal Verification',
     description: 'Must be verified by the Student Portal to access the official K-State discord.',
     type: 7,
