--- server/linkedRoles/Linked_Roles_Glitch/src/storage.js
@@ -1,9 +1,0 @@
-const store = new Map();
-
-export async function storeDiscordTokens(userId, tokens) {
-  await store.set(`discord-${userId}`, tokens);
-}
-
-export async function getDiscordTokens(userId) {
-  return store.get(`discord-${userId}`);
-}
